feat(args): Add support for specifying a directory
In the original version of this script, everything had to be in the same
directory. This adds support for adding a directory as a command line
argument to the executable (as described in the documentation) so you
don't need to copy the script to wherever the files to be re-muxed live.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>