fix(save): Correct output of converted file
I'm really messy with the naming of files in this program and it turns
out that it wasn't saving the converted files to their appropriate
directory. This fixes the output directory so it is saved appropriately.

There's a tiny fix in here that changes `recontainered` to `remuxed`.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>