fix(convert): error on zero byte files
There's a couple of different ways to think about this, but I kinda
favor crashing out if a zero byte file is created. This usually
indicates that `ffmpeg` had a problem in converting the file. Otherwise
it leaves a ton of tiny mp4 files all over the place.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>